Don't use sympy Float functions, use an opaque one with no reasoning
Sympy simplifications don't obey floating point semantics, so don't
use Sympy for this.  Keep them as is, only evaluate with the reference
implementations when all arguments are known.

This may end up getting subsumed by some other changes later, but I
wanted to understand if this was easy and it seems to be easy.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: fc4349b4fd1acc2d8b5db638a5cfb1b195d6c338
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122823